Fix warnings in Layer pages

+++ src/web/core/src/test/java/org/geoserver/web/data/layergroup/LayerGroupEditPageTest.java
@@ -172,8 +172,8 @@ public class LayerGroupEditPageTest extends LayerGroupBaseTest {
         // Ensure that the Row count is equal to the Layers in the Catalog
         Catalog catalog = getGeoServerApplication().getCatalog();
         
-        int layerCount = catalog.count(LayerInfo.class, Filter.INCLUDE);
-        int rowCount = dataView.getRowCount();
+        long layerCount = catalog.count(LayerInfo.class, Filter.INCLUDE);
+        long rowCount = dataView.getRowCount();
         
         assertEquals(layerCount, rowCount);
     }
